refactor: move controllers into cluster collector directory
