+++ autoscaler/controllers/clustercollector/configmap_test.go
@@ -1,4 +1,4 @@
-package gateway
+package clustercollector
 
 import (
 	"fmt"
