+++ autoscaler/controllers/clustercollector/gatewaydeployment_controller.go
@@ -1,11 +1,10 @@
-package controllers
+package clustercollector
 
 import (
 	"context"
 
 	"github.com/odigos-io/odigos/api/k8sconsts"
 	odigosv1 "github.com/odigos-io/odigos/api/odigos/v1alpha1"
-	predicate "github.com/odigos-io/odigos/k8sutils/pkg/predicate"
 	appsv1 "k8s.io/api/apps/v1"
 	"k8s.io/apimachinery/pkg/types"
 	ctrl "sigs.k8s.io/controller-runtime"
@@ -13,11 +12,13 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/log"
 )
 
-type GatewayDeploymentReconciler struct {
+type ClusterCollectorDeploymentReconciler struct {
 	client.Client
 }
 
-func (r *GatewayDeploymentReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
+// This controller is used to track and update the collectors group status once the cluster collector is ready to receive data.
+// We don't want to spin up components that needs to export to cluster collector and have errors and memory pressure.
+func (r *ClusterCollectorDeploymentReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
 	logger := log.FromContext(ctx)
 	logger.V(0).Info("Reconciling Deployment")
 
@@ -45,11 +46,4 @@ func (r *GatewayDeploymentReconciler) Reconcile(ctx context.Context, req ctrl.Re
 	}
 
 	return ctrl.Result{}, nil
-}
-
-func (r *GatewayDeploymentReconciler) SetupWithManager(mgr ctrl.Manager) error {
-	return ctrl.NewControllerManagedBy(mgr).
-		For(&appsv1.Deployment{}).
-		WithEventFilter(&predicate.ClusterCollectorDeploymentPredicate).
-		Complete(r)
 }
