+++ autoscaler/controllers/clustercollector/secrets.go
@@ -1,4 +1,4 @@
-package gateway
+package clustercollector
 
 import (
 	"context"
